Updates, bugfixes, README
Added nonconstant geometry support.

Fixed periodic loading support. We weren't including
the number of timesteps in the PeriodicLoadingHeader

Loosened the namelist parser so that it can read in
partial namelists to be used as a template.

Added python wrappers for periodic loading.

Added output.d and python/output.d to parse wopwop
output files and provide a programatic interface.

Added a README file with some build instructions.