Many new features
Exposed the rest of the CaseList struct to python. Also
added a namelist member to the Casename struct so the namelist
can be conviently be carried around

Added more wopwop output processing. We now read back
OASPL files, pressure time histories, spectrums and grids

We now have a module for executing wopwop as if it were part
of the library. It takes in a CaseList and returns a WopwopResult